requirements.txt
tqdm==4.55.0
pandas==1.2.0
webvtt_py==0.4.6
youtube_dl==2021.1.3
requests==2.22.0
beautifulsoup4==4.9.3
